+++ controllers/products_controller.js
@@ -17,19 +17,6 @@ async function getAllProducts (_req, res) {
 	}
 }
 
-
-// Get all product categories 
-async function getAllCategories (req, res) {
-	try{
-		const rows = await products.getAllCategories();
-		res.status(200).json(rows);
-	}
-	catch(err) {
-		console.error("Error retrieving categories:", err.message);
-		res.status(500).json({ error: "Internal server error" });
-	}
-}
-
 //Gets product with id "id"
 async function getProductWithId (req, res) {
 	const product_id = req.params.id;
@@ -49,24 +36,5 @@ async function getProductWithId (req, res) {
 
 }
 
-// Endpoint to get all products with category name
-async function getCategory (req, res) {
-	const categoryName = req.params.categoryName;
-	try{
-		const rows = await products.getCategory(categoryName);
-		if (rows.length === 0) {
-			res.status(404).json({ error: "No products found" });
-			return; 
-		}
-		res.status(200).json(rows);
-	}
-	catch(err){
-		console.error("Error retrieving category", err.message);
-		res.status(500).json({ error: "Internal server error" });
-	}
-}
-
 exports.getAllProducts = getAllProducts;
-exports.getAllCategories = getAllCategories;
 exports.getProductWithId = getProductWithId;
-exports.getCategory = getCategory;
